backend/app/services/model_service.py
from __future__ import annotations

from io import BytesIO
from hashlib import sha1

import numpy as np
from PIL import Image

from app.core.config import settings


class ModelService:
    """Prediction service with real TensorFlow inference and deterministic fallback."""

    treatment_map = {
        "Potato___Early_blight": [
            "Use certified disease-free seed tubers in the next cycle.",
            "Apply a protective fungicide program if infection spreads.",
            "Remove infected lower leaves and improve plant spacing.",
        ],
        "Potato___Late_blight": [
            "Remove and destroy infected plants immediately.",
            "Avoid overhead irrigation and keep foliage dry.",
            "Spray late-blight specific fungicide as per local guidelines.",
        ],
        "Potato___healthy": [
            "Continue balanced fertilization and irrigation schedule.",
            "Scout field weekly to detect symptoms early.",
            "Maintain crop rotation and field sanitation.",
        ],
    }

    def __init__(self) -> None:
        self.labels = [
            label.strip() for label in settings.class_labels.split(",") if label.strip()
        ]
        if not self.labels:
            self.labels = [
                "Potato___Early_blight",
                "Potato___healthy",
                "Potato___Late_blight",
            ]

        self.model = None
        self.model_ready = False
        self._load_model()

    def _load_model(self) -> None:
        try:
            import tensorflow as tf

            self.model = tf.keras.models.load_model(settings.model_path)
            # Warm up once to reduce latency on the first prediction request.
            warmup = np.zeros(
                (1, settings.image_size, settings.image_size, 3),
                dtype=np.float32,
            )
            _ = self.model.predict(warmup, verbose=0)
            self.model_ready = True
        except Exception:
            self.model = None
            self.model_ready = False

    @staticmethod
    def _normalize_scores(scores: list[float]) -> list[float]:
        total = sum(scores)
        if total <= 0:
            return [0.0 for _ in scores]
        return [score / total for score in scores]

    def _preprocess(self, file_bytes: bytes) -> np.ndarray:
        try:
            image = Image.open(BytesIO(file_bytes)).convert("RGB")
        except Exception as exc:
            raise ValueError(
                "Unable to read image. Please upload a valid image file."
            ) from exc

        image = image.resize((settings.image_size, settings.image_size))
        arr = np.array(image, dtype=np.float32) / 255.0
        return np.expand_dims(arr, axis=0)

    def _fallback_predict(self, file_bytes: bytes) -> dict:
        digest = sha1(file_bytes).hexdigest()
        idx = int(digest[:2], 16) % len(self.labels)
        base_confidence = 0.64 + ((int(digest[2:4], 16) % 25) / 100)

        raw_scores = [
            max(0.01, base_confidence - i * 0.12)
            for i in range(min(3, len(self.labels)))
        ]
        scores = self._normalize_scores(raw_scores)

        details = []
        for i, score in enumerate(scores):
            label_idx = (idx + i) % len(self.labels)
            details.append({"label": self.labels[label_idx], "score": round(score, 4)})

        top_label = details[0]["label"]
        return {
            "label": top_label,
            "confidence": details[0]["score"],
            "details": details,
            "recommendations": self.treatment_map.get(top_label, []),
            "model_ready": False,
        }

    def predict(self, file_bytes: bytes) -> dict:
        if not file_bytes:
            raise ValueError("Uploaded image is empty.")

        if not self.model_ready or self.model is None:
            return self._fallback_predict(file_bytes)

        model_input = self._preprocess(file_bytes)
        preds = self.model.predict(model_input, verbose=0)
        probs = np.asarray(preds[0], dtype=np.float32)
        k = min(3, probs.shape[0])
        top_indices = np.argpartition(-probs, k - 1)[:k]
        top_indices = top_indices[np.argsort(-probs[top_indices])]

        details = [
            {
                "label": self.labels[int(idx)]
                if int(idx) < len(self.labels)
                else f"Class_{int(idx)}",
                "score": round(float(probs[int(idx)]), 4),
            }
            for idx in top_indices
        ]

        top_label = details[0]["label"]

        return {
            "label": top_label,
            "confidence": details[0]["score"],
            "details": details,
            "recommendations": self.treatment_map.get(top_label, []),
            "model_ready": True,
        }


model_service = ModelService()
