backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.8.1
numpy==2.1.3
Pillow==11.1.0
